Add files via upload
MovieLens dataset to build a model to recommend movies to the end users. This data has been collected by the GroupLens Research Project at the University of Minnesota.we will find the similarity between each movie pair and based on that, we will recommend similar movies which are liked by the users in the past. This algorithm works similar to user-user collaborative filtering with just a little change – instead of taking the weighted sum of ratings of “user-neighbors”, we take the weighted sum of ratings of “item-neighbors”. The prediction is given by: